Move to a trait so we don't have repeated code.
